added get_apt_rating, added get_owner_rating, get_owner_rating doesnt work

+++ Solution.py
@@ -119,14 +119,44 @@ def create_tables():
     FROM Review
     GROUP BY ApartmentID;
     """
+    owner_apts_view = """
+    CREATE VIEW OwnerApartments AS
+    SELECT o.OwnerID ,a.*
+    FROM Owner o
+    JOIN Owns os ON o.OwnerID = os.OwnerID
+    JOIN Apartment a ON os.ApartmentID = a.ApartmentID;
+    """
+    # owner rating view provides a view of the average rating of each owner, using the OwnerApartments view and the Ratings view
+    owner_rating_view = """
+    CREATE VIEW OwnerRating AS
+    SELECT oa.OwnerID as OwnerID, AVG(r.AvgRating) as AvgRating
+    FROM OwnerApartments oa 
+    JOIN Ratings r ON oa.ApartmentID = r.ApartmentID
+    GROUP BY oa.OwnerID;
+    """
+    customer_reservation_view = """
+    CREATE VIEW CustomerReservations AS
+    SELECT CustomerID, COUNT(*) as Reservations
+    FROM Reservation
+    GROUPBY CustomerID
+    """
+    owner_reservation_view = """
+    CREATE VIEW OwnerReservation AS
+    SELECT oa.OwnerID, COUNT(rs.ReservationID) as ReservationCount
+    FROM OwnerApartments oa
+    JOIN Reservation rs ON rs.ApartmentID = oa.ApartmentID
+    GROUP BY oa.OwnerId
+    """
     full_query = (
-            create_customer_table
-            + create_owner_table
-            + create_apt_table
-            + create_owns_table
-            + create_reservation_table
-            + create_review_table
-            + apt_avg_rating_view
+        create_customer_table
+        + create_owner_table
+        + create_apt_table
+        + create_owns_table
+        + create_reservation_table
+        + create_review_table
+        + apt_avg_rating_view
+        + owner_apts_view
+        + owner_rating_view
     )
     conn.execute(full_query)
     conn.commit()
@@ -171,9 +201,9 @@ def add_owner(owner: Owner) -> ReturnValue:
     add_owner_query = sql.SQL(
         """
     INSERT INTO Owner (OwnerID, Name)
-    VALUES ({}, {})
+    VALUES ({owner_id}, {name})
     """
-    ).format(sql.Literal(owner_id), sql.Literal(owner_name))
+    ).format(owner_id=sql.Literal(owner_id), name=sql.Literal(owner_name))
     conn = Connector.DBConnector()
     try:
         conn.execute(add_owner_query)
@@ -189,9 +219,9 @@ def get_owner(owner_id: int) -> Owner:  # Doron
     conn = Connector.DBConnector()
     get_owner_query = sql.SQL(
         """
-    SELECT OwnerID, Name FROM Owner WHERE OwnerID = {}
+    SELECT OwnerID, Name FROM Owner WHERE OwnerID = {owner_id}
     """
-    ).format(sql.Literal(owner_id))
+    ).format(owner_id=sql.Literal(owner_id))
     try:
         num_rows, result_set = conn.execute(get_owner_query)
     except exception_list:
@@ -209,9 +239,9 @@ def delete_owner(owner_id: int) -> ReturnValue:  # Daniel
     conn = Connector.DBConnector()
     delete_owner_query = sql.SQL(
         """
-    DELETE FROM Owner WHERE OwnerID = {}
+    DELETE FROM Owner WHERE OwnerID = {owner_id}
     """
-    ).format(sql.Literal(owner_id))
+    ).format(owner_id=sql.Literal(owner_id))
     try:
         affected_lines, _ = conn.execute(delete_owner_query)
         conn.commit()
@@ -235,14 +265,14 @@ def add_apartment(apartment: Apartment) -> ReturnValue:  # Doron
     add_apartment_query = sql.SQL(
         """
     INSERT INTO Apartment (ApartmentID, Address, City, Country, Size)
-    VALUES ({}, {}, {}, {}, {})
+    VALUES ({apartment_id}, {address}, {city}, {country}, {size})
     """
     ).format(
-        sql.Literal(apartment_id),
-        sql.Literal(apartment_address),
-        sql.Literal(apartment_city),
-        sql.Literal(apartment_country),
-        sql.Literal(apartment_size),
+        apartment_id=sql.Literal(apartment_id),
+        address=sql.Literal(apartment_address),
+        city=sql.Literal(apartment_city),
+        country=sql.Literal(apartment_country),
+        size=sql.Literal(apartment_size),
     )
     try:
         conn.execute(add_apartment_query)
@@ -259,9 +289,9 @@ def get_apartment(apartment_id: int) -> Apartment:  # Daniel
     conn = Connector.DBConnector()
     get_apt_query = sql.SQL(
         """
-    SELECT * FROM Apartment WHERE ApartmentID = {}
+    SELECT * FROM Apartment WHERE ApartmentID = {apartment_id}
     """
-    ).format(sql.Literal(apartment_id))
+    ).format(apartment_id=sql.Literal(apartment_id))
     try:
         num_rows, result_set = conn.execute(get_apt_query)
     except exception_list as e:
@@ -280,9 +310,9 @@ def delete_apartment(apartment_id: int) -> ReturnValue:  # Doron
     conn = Connector.DBConnector()
     delete_apartment_query = sql.SQL(
         """
-    DELETE FROM Apartment WHERE ApartmentID = {}
+    DELETE FROM Apartment WHERE ApartmentID = {apartment_id}
     """
-    ).format(sql.Literal(apartment_id))
+    ).format(apartment_id=sql.Literal(apartment_id))
     try:
         rows_affected, _ = conn.execute(delete_apartment_query)
     except exception_list as e:
@@ -304,9 +334,9 @@ def add_customer(customer: Customer) -> ReturnValue:  # Daniel
     add_customer_query = sql.SQL(
         """
     INSERT INTO Customer (CustomerID, Name)
-    VALUES({}, {})
+    VALUES({customer_id}, {customer_name})
     """
-    ).format(sql.Literal(customer_id), sql.Literal(customer_name))
+    ).format(customer_id=sql.Literal(customer_id), customer_name=sql.Literal(customer_name))
     try:
         conn.execute(add_customer_query)
         conn.commit()
@@ -322,9 +352,9 @@ def get_customer(customer_id: int) -> Customer:  # Doron
     conn = Connector.DBConnector()
     get_customer_query = sql.SQL(
         """
-    SELECT CustomerID, Name FROM Customer WHERE CustomerID = {}
+    SELECT CustomerID, Name FROM Customer WHERE CustomerID = {customer_id}
     """
-    ).format(sql.Literal(customer_id))
+    ).format(customer_id=sql.Literal(customer_id))
     rows, result_set = conn.execute(get_customer_query)
     conn.close()
     if rows < 1:
@@ -337,9 +367,9 @@ def delete_customer(customer_id: int) -> ReturnValue:  # Daniel
     conn = Connector.DBConnector()
     delete_customer_query = sql.SQL(
         """
-    DELETE FROM Customer WHERE CustomerID = {}
+    DELETE FROM Customer WHERE CustomerID = {customer_id}
     """
-    ).format(sql.Literal(customer_id))
+    ).format(customer_id=sql.Literal(customer_id))
     try:
         rows_affected, _ = conn.execute(delete_customer_query)
     except exception_list as e:
@@ -353,33 +383,30 @@ def delete_customer(customer_id: int) -> ReturnValue:  # Daniel
 
 # Customer made a reservation of apartment from start_date to end_date and paid total_price
 def customer_made_reservation(
-        customer_id: int,
-        apartment_id: int,
-        start_date: date,
-        end_date: date,
-        total_price: float,
+    customer_id: int,
+    apartment_id: int,
+    start_date: date,
+    end_date: date,
+    total_price: float,
 ) -> ReturnValue:  # Doron
     conn = Connector.DBConnector()
-    customer_made_reservation_query = sql.SQL("""
-                            INSERT INTO Reservations(CustomerID, ApartmentID, StartDate, EndDate, Price) 
-                                VALUES({customer_id_t}, {apartment_id_t}, {start_date_t}, {end_date_t}, {total_price_t})
-                                WHERE NOT EXISTS (
-                                    SELECT * FROM Reservations 
-                                    WHERE apartment_id = {apartment_id_t} 
-                                    AND (
-                                        {start_date_t} BETWEEN start_date AND end_date OR
-                                        {end_date_t} BETWEEN start_date AND end_date OR
-                                        start_date BETWEEN {start_date_t} AND {end_date_t} OR
-                                        end_date BETWEEN {start_date_t} AND {end_date_t}
-                                    )
-                                );
-                                """).format(customer_id_t=sql.Literal(customer_id),
-                                            apartment_id_t=sql.Literal(apartment_id),
-                                            start_date_t=sql.Literal(start_date),
-                                            end_date_t=sql.Literal(end_date),
-                                            total_price_t=sql.Literal(total_price)
-                                            )
-
+    customer_made_reservation_query = sql.SQL(
+        """
+    INSERT INTO Reservation (CustomerID, ApartmentID, StartDate, EndDate, Price) 
+    SELECT {customer_id}, {apartment_id}, {start_date}, {end_date}, {total_price}
+    WHERE NOT EXISTS (
+        SELECT * FROM Reservation
+        WHERE ApartmentId = {apartment_id} 
+        AND StartDate <= {end_date} AND EndDate >= {start_date}
+    );
+    """
+    ).format(
+        customer_id=sql.Literal(customer_id),
+        apartment_id=sql.Literal(apartment_id),
+        start_date=sql.Literal(start_date),
+        end_date=sql.Literal(end_date),
+        total_price=sql.Literal(total_price),
+    )
     try:
         rows_affected, _ = conn.execute(customer_made_reservation_query)
         if rows_affected == 0:
@@ -395,15 +422,15 @@ def customer_made_reservation(
 
 # Remove a reservation from the database.
 def customer_cancelled_reservation(
-        customer_id: int, apartment_id: int, start_date: date
+    customer_id: int, apartment_id: int, start_date: date
 ) -> ReturnValue:  # Daniel
     conn = Connector.DBConnector()
     customer_cancelled_reservation_query = sql.SQL(
         """
-    DELETE FROM Reservation WHERE CustomerID = {} AND ApartmentID = {} AND StartDate = {}
+    DELETE FROM Reservation WHERE CustomerID = {customer_id} AND ApartmentID = {apartment_id} AND StartDate = {start_date}
     """
     ).format(
-        sql.Literal(customer_id), sql.Literal(apartment_id), sql.Literal(start_date)
+        customer_id=sql.Literal(customer_id), apartment_id=sql.Literal(apartment_id), start_date=sql.Literal(start_date)
     )
     try:
         rows_affected, _ = conn.execute(customer_cancelled_reservation_query)
@@ -419,33 +446,30 @@ def customer_cancelled_reservation(
 
 # Customer reviewed apartment on date review_date and gave it rating stars, with text review_text.
 def customer_reviewed_apartment(
-        customer_id: int,
-        apartment_id: int,
-        review_date: date,
-        rating: int,
-        review_text: str,
+    customer_id: int,
+    apartment_id: int,
+    review_date: date,
+    rating: int,
+    review_text: str,
 ) -> ReturnValue:  # Doron
     conn = Connector.DBConnector()
     customer_reviewed_apartment_query = sql.SQL(
         """
     Insert into Review (CustomerID, ApartmentID, ReviewDate, Rating, ReviewText)
-    SELECT {}, {}, {}, {}, {}
+    SELECT {customer_id}, {apartment_id}, {review_date}, {rating}, {review_text}
     WHERE EXISTS (
         SELECT * FROM Reservation
-        WHERE CustomerID = {}
-        AND ApartmentID = {}
-        AND (EndDate < {})
+        WHERE CustomerID = {customer_id}
+        AND ApartmentID = {apartment_id}
+        AND (EndDate < {review_date})
     );
     """
     ).format(
-        sql.Literal(customer_id),
-        sql.Literal(apartment_id),
-        sql.Literal(review_date),
-        sql.Literal(rating),
-        sql.Literal(review_text),
-        sql.Literal(customer_id),
-        sql.Literal(apartment_id),
-        sql.Literal(review_date),
+        customer_id=sql.Literal(customer_id),
+        apartment_id=sql.Literal(apartment_id),
+        review_date=sql.Literal(review_date),
+        rating=sql.Literal(rating),
+        review_text=sql.Literal(review_text),
     )
     try:
         rows_affected, _ = conn.execute(customer_reviewed_apartment_query)
@@ -461,31 +485,28 @@ def customer_reviewed_apartment(
 
 # Customer decided to update their review of apartment on update_date and changed his rating to new_rating and the review text to new_text
 def customer_updated_review(
-        customer_id: int,
-        apartment_id: int,
-        update_date: date,
-        new_rating: int,
-        new_text: str,
+    customer_id: int,
+    apartment_id: int,
+    update_date: date,
+    new_rating: int,
+    new_text: str,
 ) -> ReturnValue:  # Daniel
     conn = Connector.DBConnector()
     customer_updated_review_query = sql.SQL(
         """
     UPDATE Review 
-    SET reviewdate = {}, rating = {}, reviewtext = {}
-    WHERE CustomerID = {} AND ApartmentID = {} AND {} > (
+    SET reviewdate = {update_date}, rating = {new_rating}, reviewtext = {new_text}
+    WHERE CustomerID = {customer_id} AND ApartmentID = {apartment_id} AND {update_date} > (
         SELECT reviewdate FROM review
-        WHERE CustomerID = {} AND ApartmentID = {}
+        WHERE CustomerID = {customer_id} AND ApartmentID = {apartment_id}
     )
     """
     ).format(
-        sql.Literal(update_date),
-        sql.Literal(new_rating),
-        sql.Literal(new_text),
-        sql.Literal(customer_id),
-        sql.Literal(apartment_id),
-        sql.Literal(update_date),
-        sql.Literal(customer_id),
-        sql.Literal(apartment_id),
+        update_date=sql.Literal(update_date),
+        new_rating=sql.Literal(new_rating),
+        new_text=sql.Literal(new_text),
+        customer_id=sql.Literal(customer_id),
+        apartment_id=sql.Literal(apartment_id),
     )
     try:
         rows_affected, _ = conn.execute(customer_updated_review_query)
@@ -505,13 +526,11 @@ def owner_owns_apartment(owner_id: int, apartment_id: int) -> ReturnValue:  # Do
     owner_owns_apartment_query = sql.SQL(
         """
     INSERT INTO Owns (OwnerID, ApartmentID)
-    VALUES ({}, {})
+    VALUES ({owner_id}, {apartment_id})
     """
     ).format(
-        sql.Literal(owner_id),
-        sql.Literal(apartment_id),
-        sql.Literal(apartment_id),
-        sql.Literal(owner_id),
+        owner_id=sql.Literal(owner_id),
+        apartment_id=sql.Literal(apartment_id),
     )
     try:
         conn.execute(owner_owns_apartment_query)
@@ -529,9 +548,9 @@ def owner_drops_apartment(owner_id: int, apartment_id: int) -> ReturnValue:  # D
     owner_drops_apartment_query = sql.SQL(
         """
     DELETE FROM Owns
-    WHERE OwnerID = {} AND ApartmentID = {}
+    WHERE OwnerID = {owner_id} AND ApartmentID = {apartment_id}
     """
-    ).format(sql.Literal(owner_id), sql.Literal(apartment_id))
+    ).format(owner_id=sql.Literal(owner_id), apartment_id=sql.Literal(apartment_id))
     try:
         rows_affected, _ = conn.execute(owner_drops_apartment_query)
     except exception_list as e:
@@ -552,9 +571,9 @@ def get_apartment_owner(apartment_id: int) -> Owner:  # Doron
     SELECT o.OwnerID, o.Name
     FROM Owner o
     JOIN Owns os ON o.OwnerID = os.OwnerID
-    WHERE os.ApartmentID = {}
+    WHERE os.ApartmentID = {apartment_id}
     """
-    ).format(sql.Literal(apartment_id))
+    ).format(apartment_id=sql.Literal(apartment_id))
     try:
         num_rows, result_set = conn.execute(get_apartment_owner_query)
     except exception_list as e:
@@ -569,13 +588,11 @@ def get_apartment_owner(apartment_id: int) -> Owner:  # Doron
 # Get a list of all apartments owned by owner.
 def get_owner_apartments(owner_id: int) -> List[Apartment]:  # Daniel
     conn = Connector.DBConnector()
-    get_owner_apartments_query = f"""
-    SELECT a.*
-    FROM Owner o
-    JOIN Owns os ON o.OwnerID = os.OwnerID
-    JOIN Apartment a ON os.ApartmentID = a.ApartmentID
-    WHERE o.OwnerID = {owner_id}
-    """
+    get_owner_apartments_query = sql.SQL("""
+    SELECT *
+    FROM OwnerApartments    
+    WHERE OwnerID = {owner_id}                      
+    """).format(owner_id=sql.Literal(owner_id))
     try:
         num_apts, apts_data = conn.execute(get_owner_apartments_query)
     except exception_list as e:
@@ -596,8 +613,9 @@ def get_apartment_rating(apartment_id: int) -> float:
     conn = Connector.DBConnector()
     get_apartment_rating_query = sql.SQL(
         """
-    SELECT AvgRating FROM Ratings WHERE ApartmentID = {}
-    """).format(sql.Literal(apartment_id))
+    SELECT AvgRating FROM Ratings WHERE ApartmentID = {apartment_id}
+    """
+    ).format(apartment_id=sql.Literal(apartment_id))
     try:
         rows, result = conn.execute(get_apartment_rating_query)
     except exception_list as e:
@@ -611,8 +629,20 @@ def get_apartment_rating(apartment_id: int) -> float:
 
 # Get the average of averages of ratings from all reviews of apartments owned by owner.
 def get_owner_rating(owner_id: int) -> float:
-    # TODO: implement
-    pass
+    get_owner_rating_query = sql.SQL(
+    """
+    SELECT AvgRating FROM OwnerRating WHERE OwnerID = {owner_id}
+    """
+    ).format(owner_id=sql.Literal(owner_id))
+    try:
+        rows, result = conn.execute(get_owner_rating_query)
+    except exception_list as e:
+        conn.close()
+        return handle_errors(e)
+    conn.close()
+    if rows < 1:
+        return 0
+    return result[0]["AvgRating"]
 
 
 # Get the customer that made the most reservations.
@@ -665,6 +695,7 @@ def get_apartment_recommendation(customer_id: int) -> List[Tuple[Apartment, floa
 
 # Utility functions:
 
+
 def create_owner_from_response(res: ResultSetDict) -> Owner:
     try:
         return Owner(res["OwnerID"], res["Name"])
@@ -697,9 +728,17 @@ def handle_errors(e: DatabaseException):
     # print(f"handling error: {e_name}")
     if e_name == "UNIQUE_VIOLATION":
         return ReturnValue["ALREADY_EXISTS"]
-    elif (e_name == "NOT_NULL_VIOLATION" or e_name == "FOREIGN_KEY_VIOLATION" or e_name == "CHECK_VIOLATION"):
+    elif (
+        e_name == "NOT_NULL_VIOLATION"
+        or e_name == "FOREIGN_KEY_VIOLATION"
+        or e_name == "CHECK_VIOLATION"
+    ):
         return ReturnValue["BAD_PARAMS"]
-    elif (e_name == "database_ini_ERROR" or e_name == "UNKNOWN_ERROR" or e_name == "ConnectionInvalid"):
+    elif (
+        e_name == "database_ini_ERROR"
+        or e_name == "UNKNOWN_ERROR"
+        or e_name == "ConnectionInvalid"
+    ):
         return ReturnValue["ERROR"]
